Fix warning (comparison signed and unsigned)

+++ opencog/learning/moses/metapopulation/metapopulation.cc
@@ -117,7 +117,7 @@ scored_combo_tree_ptr_set::const_iterator metapopulation::select_exemplar()
     if (size() == 1) {
         scored_combo_tree_ptr_set::const_iterator selex = _scored_trees.cbegin();
         if(_params.revisit < 0 or
-           (_params.revisit + 1 > _visited_exemplars[*selex])) // not enough visited
+           (_params.revisit + 1 > (int)_visited_exemplars[*selex])) // not enough visited
             _visited_exemplars[*selex]++;
         else selex = _scored_trees.cend();    // enough visited
 
@@ -139,7 +139,7 @@ scored_combo_tree_ptr_set::const_iterator metapopulation::select_exemplar()
 
         // Skip exemplars that have been visited enough
         if (_params.revisit < 0 or
-            (_params.revisit + 1 > _visited_exemplars[bsct])) {
+            (_params.revisit + 1 > (int)_visited_exemplars[bsct])) {
             probs.push_back(sc);
             found_exemplar = true;
             if (highest_score < sc) highest_score = sc;
